+++ tests/Fixtures/InvalidTrailingSlash/InvalidTrailingSlashController.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\Fixtures\InvalidTrailingSlash;
+
+use Karewan\KnRoute\Attributes\Get;
+
+class InvalidTrailingSlashController
+{
+	#[Get('/invalid/')]
+	public function invalid(): void {}
+}